Create whooping_crane_analysis
This is a script that creates a given geometry, in this case a whooping crane, pre-processes the geometry, solves the eigenproblem, displays the eigenvectors, solves a static loading case and then solves a transient response case. Currently the transient response part is really slow due to 'interp1' in matlab being slow.